Writing common ML benchmark class for tabular collection
